+++ nyalcf_cli/bin/nyalcf_cli.dart
@@ -3,7 +3,9 @@ import 'package:nyalcf_cli/arguments.dart';
 import 'package:nyalcf_core/storages/injector.dart';
 import 'package:nyalcf_core/utils/logger.dart';
 import 'package:nyalcf_core/utils/path_provider.dart';
-import 'package:nyalcf_core/utils/universe.dart';
+import 'package:nyalcf_inject/nyalcf_inject.dart';
+
+final version = '0.1.0';
 
 ArgParser buildParser() {
   return Arguments.all;
@@ -16,12 +18,15 @@ void printUsage(ArgParser argParser) {
 
 void main(List<String> arguments) async {
   /// 加载环境
-  await Universe.loadUniverse();
+  // await Universe.loadUniverse();
+  setAppendInfo('CLI v$version');
 
   /// 初始化配置文件等
   await PathProvider.loadSyncPath();
   await StoragesInjector.init();
   await Logger.init();
+
+  Logger.debug('Append info has been set: $appendInfo');
 
   final ArgParser argParser = buildParser();
   try {
@@ -39,7 +44,7 @@ void main(List<String> arguments) async {
     }
 
     if (results.wasParsed('version')) {
-      print('Nya LoCyanFrp! core version: ${Universe.appVersion}');
+      TODO: print('Nya LoCyanFrp! CLI version: $version');
       return;
     }
     if (results.wasParsed('verbose')) {
