+++ nyalcf_cli/pubspec.yaml
@@ -10,6 +10,8 @@ environment:
 # Add regular dependencies here.
 dependencies:
   args: ^2.4.2
+  nyalcf_inject:
+    path: ../nyalcf_inject
   nyalcf_core:
     path: ../nyalcf_core
