+++ nyalcf_gui/nyalcf_core_ui/lib/utils/universe.dart
