+++ nyalcf_gui/nyalcf_core_ui/pubspec.yaml
@@ -25,17 +25,18 @@ dependencies:
   flutter:
     sdk: flutter
 
-  nyalcf_inject:
-    path: ../nyalcf_inject
   nyalcf_core:
     path: ../../nyalcf_core
   nyalcf_ui:
     path: ../nyalcf_ui
+  nyalcf_inject_ui:
+    path: ../nyalcf_inject_ui
   get: ^4.6.6
   highlight: ^0.7.0
   url_launcher: ^6.2.6
   dio: ^5.5.0+1
   shared_preferences: ^2.2.3
+  package_info_plus: ^7.0.0
 
 dev_dependencies:
   flutter_test:
