+++ nyalcf_gui/nyalcf_inject_ui/analysis_options.yaml
