+++ nyalcf_gui/nyalcf_inject_ui/pubspec.yaml
@@ -1,4 +1,4 @@
-name: nyalcf_inject
+name: nyalcf_inject_ui
 description: "Nya LoCyanFrp! communication injector."
 version: 1.0.3
 homepage: https://nyalcf.1l1.icu
