+++ nyalcf_gui/nyalcf_ui/lib/views/auth/tokenmode.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 import 'package:nyalcf_core_ui/storages/prefs/token_mode_prefs.dart';
-import 'package:nyalcf_inject/nyalcf_inject.dart';
+import 'package:nyalcf_inject_ui/nyalcf_inject_ui.dart';
 import 'package:nyalcf_ui/models/appbar_actions.dart';
 import 'package:nyalcf_ui/models/floating_action_button.dart';
