+++ nyalcf_gui/nyalcf_ui/lib/views/home.dart
@@ -11,7 +11,7 @@ import 'package:nyalcf_core_ui/utils/frpc/startup_loader.dart';
 import 'package:nyalcf_core/utils/logger.dart';
 import 'package:nyalcf_core/network/dio/auth/auth.dart';
 import 'package:nyalcf_core_ui/utils/proxies_getter.dart';
-import 'package:nyalcf_inject/nyalcf_inject.dart';
+import 'package:nyalcf_inject_ui/nyalcf_inject_ui.dart';
 import 'package:nyalcf_ui/models/appbar_actions.dart';
 import 'package:nyalcf_ui/models/floating_action_button.dart';
