+++ nyalcf_gui/nyalcf_ui/lib/views/license.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import 'package:nyalcf_core/utils/universe.dart';
-import 'package:nyalcf_inject/nyalcf_inject.dart';
+import 'package:nyalcf_core_ui/utils/universe.dart';
+import 'package:nyalcf_inject_ui/nyalcf_inject_ui.dart';
 import 'package:nyalcf_ui/models/appbar_actions.dart';
 import 'package:nyalcf_ui/models/floating_action_button.dart';
