+++ nyalcf_gui/nyalcf_ui/lib/views/panel/proxies.dart
@@ -5,7 +5,7 @@ import 'package:nyalcf_core_ui/utils/widget/after_layout.dart';
 import 'package:nyalcf_ui/controllers/proxies_controller.dart';
 import 'package:nyalcf_ui/controllers/user_controller.dart';
 import 'package:nyalcf_core/storages/stores/proxies_storage.dart';
-import 'package:nyalcf_inject/nyalcf_inject.dart';
+import 'package:nyalcf_inject_ui/nyalcf_inject_ui.dart';
 import 'package:nyalcf_ui/models/account_dialog.dart';
 import 'package:nyalcf_ui/models/appbar_actions.dart';
 import 'package:nyalcf_ui/models/drawer.dart';
