+++ nyalcf_gui/nyalcf_ui/lib/views/setting/injector.dart
@@ -3,7 +3,7 @@ import 'package:get/get.dart';
 
 import 'package:nyalcf_ui/controllers/frpc_setting_controller.dart';
 import 'package:nyalcf_ui/controllers/launcher_setting_controller.dart';
-import 'package:nyalcf_inject/nyalcf_inject.dart';
+import 'package:nyalcf_inject_ui/nyalcf_inject_ui.dart';
 import 'package:nyalcf_ui/models/appbar_actions.dart';
 import 'package:nyalcf_ui/models/floating_action_button.dart';
 import 'package:nyalcf_ui/views/setting/frpc_setting.dart';
