+++ nyalcf_gui/nyalcf_ui/lib/views/tokenmode/panel.dart
@@ -6,7 +6,7 @@ import 'package:nyalcf_ui/controllers/frpc_controller.dart';
 import 'package:nyalcf_core_ui/storages/prefs/token_mode_prefs.dart';
 import 'package:nyalcf_core/utils/frpc/path_provider.dart';
 import 'package:nyalcf_core_ui/utils/frpc/process_manager.dart';
-import 'package:nyalcf_inject/nyalcf_inject.dart';
+import 'package:nyalcf_inject_ui/nyalcf_inject_ui.dart';
 import 'package:nyalcf_ui/models/appbar_actions.dart';
 import 'package:nyalcf_ui/models/floating_action_button.dart';
 import 'package:nyalcf_ui/models/process_list_dialog.dart';
