+++ nyalcf_gui/nyalcf_ui/pubspec.yaml
@@ -12,10 +12,10 @@ dependencies:
   flutter:
     sdk: flutter
 
-  nyalcf_inject:
-    path: ../nyalcf_inject
   nyalcf_core:
     path: ../../nyalcf_core
+  nyalcf_inject_ui:
+    path: ../nyalcf_inject_ui
   nyalcf_core_ui:
     path: ../nyalcf_core_ui
