+++ nyalcf_gui/pubspec.yaml
@@ -32,7 +32,7 @@ dependencies:
     sdk: flutter
 
   nyalcf_inject:
-    path: ./nyalcf_inject
+    path: ../nyalcf_inject
   nyalcf_core:
     path: ../nyalcf_core
   nyalcf_core_ui:
