+++ nyalcf_inject/lib/nyalcf_inject.dart
@@ -1,0 +1,5 @@
+String appendInfo = "Unknown";
+
+void setAppendInfo(value) {
+  appendInfo = value;
+}
